requirements.txt
Flask
flask-cors
werkzeug
pandas
openpyxl
pdfplumber
tabula-py
camelot-py[cv]
